Updated and optimized the model and Bridge
Proper optimization and usage of mediapipeai for ios and android with gpu implementation. Usage of singleton and lazy loading for proper meomery management and to avoid the leakage. Added stats and warmup to activate the model and tested on iphone 14 wont work. but takes 3 gb meomry on simulator iphone 16 pro